+++ src/somesy/julia/writer.py
@@ -59,7 +59,7 @@ class Julia(ProjectMetadataWriter):
         try:
             return Person.from_name_email_string(person)
         except (ValueError, AttributeError):
-            logger.warning(f"Cannot convert {person} to Person object, trying Entity.")
+            logger.info(f"Cannot convert {person} to Person object, trying Entity.")
 
         try:
             return Entity.from_name_email_string(person)