+++ src/somesy/rust/writer.py
@@ -97,7 +97,7 @@ class Rust(ProjectMetadataWriter):
         try:
             return Person.from_name_email_string(person)
         except (ValueError, AttributeError):
-            logger.warning(f"Cannot convert {person} to Person object, trying Entity.")
+            logger.info(f"Cannot convert {person} to Person object, trying Entity.")
 
         try:
             return Entity.from_name_email_string(person)